Use var for long factory name
