+++ backend/common/src/main/java/ai/verta/modeldb/common/futures/InternalFutureWithDefaultExecutor.java
@@ -1,7 +1,5 @@
 package ai.verta.modeldb.common.futures;
 
-import com.google.common.util.concurrent.ListenableFuture;
-import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.Executor;
@@ -115,50 +113,5 @@ public class InternalFutureWithDefaultExecutor<T> extends InternalFuture<T> {
   public <U> InternalFutureWithDefaultExecutor<U> thenSupply(
       Supplier<InternalFuture<U>> supplier, Executor executor) {
     return super.thenSupply(supplier, executor).thenWithExecutor(defaultExecutor);
-  }
-
-  public static class FactoryWithExecutor {
-
-    private Executor executor;
-
-    FactoryWithExecutor(Executor executor) {
-      Objects.requireNonNull(executor, "Valid executor required");
-      this.executor = executor;
-    }
-
-    public <U> InternalFutureWithDefaultExecutor<U> completedInternalFuture(U value) {
-      return InternalFuture.completedInternalFuture(value).thenWithExecutor(executor);
-    }
-
-    public <U> InternalFutureWithDefaultExecutor<U> from(CompletionStage<U> stage) {
-      return InternalFuture.from(stage).thenWithExecutor(executor);
-    }
-
-    public <U> InternalFutureWithDefaultExecutor<U> from(ListenableFuture<U> listenableFuture) {
-      return InternalFuture.from(listenableFuture, executor).thenWithExecutor(executor);
-    }
-
-    public InternalFutureWithDefaultExecutor<Void> runAsync(Runnable runnable) {
-      return InternalFuture.runAsync(runnable, executor).thenWithExecutor(executor);
-    }
-
-    public <U> InternalFutureWithDefaultExecutor<List<U>> sequence(
-        final List<InternalFuture<U>> futures) {
-      return InternalFuture.sequence(futures, executor).thenWithExecutor(executor);
-    }
-
-    public <U> InternalFutureWithDefaultExecutor<U> supplyAsync(Supplier<U> supplier) {
-      return InternalFuture.supplyAsync(supplier, executor).thenWithExecutor(executor);
-    }
-
-    public <U> InternalFuture<U> failedStage(Throwable ex) {
-      return InternalFuture.<U>failedStage(ex).thenWithExecutor(executor);
-    }
-
-    public <U> InternalFutureWithDefaultExecutor<U> retriableStage(
-        Supplier<InternalFuture<U>> supplier, Function<Throwable, Boolean> retryChecker) {
-      return InternalFuture.retriableStage(supplier, retryChecker, executor)
-          .thenWithExecutor(executor);
-    }
   }
 }
