+++ backend/common/src/main/java/ai/verta/modeldb/common/reconcilers/SendEventsWithCleanUp.java
@@ -5,7 +5,7 @@ import ai.verta.modeldb.common.connections.UAC;
 import ai.verta.modeldb.common.event.FutureEventDAO;
 import ai.verta.modeldb.common.futures.FutureJdbi;
 import ai.verta.modeldb.common.futures.InternalFuture;
-import ai.verta.modeldb.common.futures.InternalFutureWithDefaultExecutor;
+import ai.verta.modeldb.common.futures.InternalFutureWithDefaultExecutorFactory;
 import ai.verta.uac.CreateEventRequest;
 import com.google.protobuf.Any;
 import com.google.protobuf.Value;
@@ -19,7 +19,7 @@ public class SendEventsWithCleanUp extends Reconciler<CreateEventRequest> {
   private final UAC uac;
   private final FutureEventDAO futureEventDAO;
 
-  private final InternalFutureWithDefaultExecutor.FactoryWithExecutor futureFactory;
+  private final InternalFutureWithDefaultExecutorFactory futureFactory;
 
   public SendEventsWithCleanUp(
       ReconcilerConfig config,
